refactor parse
